requirements.txt
wheel==0.23.0
docopt==0.6.1
pyserial
pymavlink
schedule